Create VCF index in the current directory
There are some oddities with the way Terra saves workflow file outputs
to the data tables. Say the input VCF path to the IndexVcf task is
"gs://foo/bar/baz.vcf.gz". When the task runs, the VCF will be localized
to "/cromwell_root/foo/bar/baz.vcf.gz" and running `tabix` will create
the index at "/cromwell_root/foo/bar/baz.vcf.gz.tbi". If the output file
of the task is given as `vcf + ".tbi"`, the index will delocalized to
"gs://{path_to_task_output_bucket}/foo/bar/baz.vcf.gz.tbi", but the path
saved to the Terra data table will be "gs://foo/bar/baz.vcf.gz.tbi",
which what the output file variable holds.

It seems more sensible that all paths saved to the Terra data table
properly point to the delocalized paths, but perhaps it is also logical
to save the path as it is given in the output variables.

To avoid this mess, the task has been modified to generate the index in
the current directory and then delocalize that file.