Add workflow to move old filters to INFO
BOTHSIDES_SUPPORT, HIGH_SR_BACKGROUND and PESR_GT_OVERDISPERSION are
values that were used in the VCF FILTER field in pervious versions of
GATK-SV, but are now placed in the INFO field. The MoveVcfFiltersToInfo
WDL in this commit updates older VCFs and moves those FILTER values to
the INFO field.