Scale resources for counting de novos.
The current bottleneck in the EstimateDeNovos workflow is the
CountDeNovos task and because the task is scattered by chromosome, the
runtime of each job in the scatter is variable. A simple way to speed up
the task is to scale the CPUs and memory by the number of genotypes
given as input. This is somewhat fragile because the size of the files
is used to estimate the number of genotypes, but the file size can
change if the number of columns increases. There are some hardcoded
limits to prevent extreme resource usage.