VisualizeCnvs: fix batch ordering in sharding
The order of the batches in the manifests produced by ShardVariants must
be the same within each shard otherwise the batche IDs could be linked
to the incorrect files. Also, ShardVariants must output a list of
batches in each shard so that the batches can be linked to their data
files.