VisualizeCnvs: write extracted directory to map
In MakePlots, the subset bincov archives are localized to their
execution directory paths and then extracted to the current working
directory, but the paths in the bincov map are the localization paths so
the plotting script cannot find the extracted files. We write the paths
of the extracted directories to the bincov map to fix the issue.