visualize_cnvs: add workaround for duplicate samples
In the case where samples are assigned to multiple batches in error, a
sample can be a carrier in one batch and a background sample in another
batch, for the purposes of plotting. When the bincov matrices for these
batches are merged, the ID of the sample in multiple batches will be
modified with a suffix to prevent duplicate column names in the merged
matrix. Changing the column names violates the assumption that carriers
are in the merged bincov matrix and causes an error. Because the
carrier-batch assignments are known, prior to merging bincov matrices,
we can remove any samples that are present in a batch they were not
assigned to.

It is also possible that a sample becomes a background sample for
multiple batches, but we don't care if the IDs of background samples are
modified. Background samples are just defined as non-carrier samples. Of
course there is still the issue of duplicate samples being plotted, but
if there are enough duplicate samples that the plot is messed up, the
whole callset is probably ruined.