SplitVcfByContig: fix check for contig
The check for the contig in the VCF using `grep` used an incorrect
option and regular expression. The `-F` option to treat the pattern as a
fixed string was used even though the pattern was a regular expression
and the regular expression was incorrect anyways because the output of
`bcftools stats` is a three-column table with the sequence name in the
first column and the regular expression tried to match the whole line.
A `cut` is now used to isolate the sequence name and the pattern search
is done with fixed strings and complete word to avoid things like "chr1"
matching "chr11".