Add workflow to add contig header lines to VCF
The SVAnnotate tool in GATK possibly throws an exception when the
annotation files have contigs which are not in the header of the VCF
being annotated. This workflow will be used to add header contig lines
for chr1-22,X,Y to a VCF that needs to be annotated with SVAnnotate.