benchmark_denovo: fix not de novo bug
The script assumes that the test set will have at least one not de novo
call and crashes if the assumption is violated. We add some checks to
make sure that the input sets have the data we expect.

Also update naming of "CleanVcf" to reflect the fact that the input VCF
does not actually have to be from CleanVcf.