benchmark_denovo: clarify numbers in overlap plot
The plot was misleading because it treated the de novo pipeline output
as the universe of calls in the input VCF for comparison between the
two, but used the actual set of calls in the input VCF when comparing it
to the truth calls. The overlap plot now reflects the actual counts.