visualize CNVs: fix bug with selecting spaced intervals
When the total number of gaps between the intervals to plot can be
evenly distributed, the algorithm will erroneously add `NA` indicies to
the end of the vector. For example, if the number of intervals to plot
is 26 and the total number of intervals in the coverage matrix being
plotted is 601, then the total number of gaps is 575 which can be evenly
distributed between the 26 intervals. The algorithm attempts to assign
extra gaps from right to left, but assumes there is always at least one
extra gap. When there aren't any extra gaps, the use of the colon
operator to pick which gaps to increment creates a sequence with
indicies past the end of the gap vector, resulting in NA indicies.

The bug is fixed by using `seq_len` to generate the vector of indicies
to increment so that if there aren't any remaining gaps, an empty vector
is made and nothing incremented. This changes the distribution of
remaining gaps to proceed left to right.