benchmark: fix subsetting VCFs
All of the VCFs need to be subset for the benchmarking to be meaningful.
The starting VCF (input to the de novo pipeline) needs to be subset to
SSC samples. Then the truth and evaluation VCFs need to be subset to
the samples remaining in the starting VCF so the samples in the all the
VCFs will be consistent.