benchmark: fix ordering contig header lines
Using the `--fai` option of `bcftools reheader` will only change the set
of the contig headers in the VCF/BCF to match the contigs in the index.
It will not change the order. Add a step to generate the entire new
header with the correct contig lines order and use that to reheader the
VCF.