Add script to set VCF SVLEN field
SVConcordance requires an INFO/SVLEN field for all sites and most SVs in
a VCF can be annotated with SVLEN if the INFO/END field is present. For
DUP, DEL, and INV, SVLEN is END - POS + 1. For INS, SVLEN cannot be
computed from INFO/END, but INS sites should be matched only with POS,
which can be accomplished by setting the SVLEN for all INS to an
arbitrary number and giving a 0% minimum size similarity cutoff to
SVConcordance. Other SVTYPEs are complicated and ignored.