Add script to set VCF ALGORITHMS field
SVConcordance requires an INFO/ALGORITHMS field for all sites as it will
compare depth-only site to each other, PE/SR-only sites to each
other, and mixed algorithms sites to each other. When a VCF does not
have the annotation, it is useful to still be able to match SVs using
the usual reciprocal overlap, size similarity, and breakend window. In
order for SVConcordance to match using these criteria and ignore
algorithms, it is necessary to annotate all sites with the same
algorithm so it is effectively ignored.