VisualizeGDs: refactor GD prediction script
- refactor to wrap more code into functions
- implement anchoring for terminal non-NAHR GDs
- implement a hard-coded minimum window size for non-NAHR GDs

For terminal type non-NAHR GDs, only check the first or last window in
the GD region for a shifted read-depth ratio.

The sliding window size for non-NAHR GD regions is determined by the
size of region so small regions can result in very small windows which
end up being too noisy to reliably check for shifted read depth. We set
a minimum number of bins for the window size to prevent the window from
becoming too small.