VisualizeGDs: handle samples with missing sex info
When a sample is missing a sex assignment, the expected read depth ratio
for chrX and chrY for those samples will be missing which causes all
comparisions to those values to also be missing (due to the way R deals
with NA). That makes the read depth ratio shift filter return NA for all
those samples, which, when used to index into the vector of sample
identifiers to retrieve putative carriers, produces NA values. The
function `predict_gd_carriers` should not return missing values as all
sample IDs returned by the function are assumed to be valid. Also, if a
sample is missing a ploidy value for the sex chromosomes, then CNVs
cannot be called for that sample using read depth.

The expected read depth ratio is now checked for missingness.